RIScraper/requirements.txt
atomicwrites==1.4.1
attrs==22.1.0
certifi==2022.6.15
charset-normalizer==2.1.1
colorama==0.4.5
greenlet==1.1.2
html5lib==1.1
idna==3.3
iniconfig==1.1.1
lxml==4.9.1
mechanize==0.4.8
nameparser==1.1.1
packaging==21.3
pluggy==1.0.0
psycopg2==2.9.3
py==1.11.0
pycodestyle==2.9.1
pyflakes==2.5.0
pyparsing==3.0.9
pytest==7.1.2
PyYAML==6.0
requests==2.28.1
six==1.16.0
SQLAlchemy==1.4.40
tomli==2.0.1
webencodings==0.5.1
structlog==22.1.0